insane hack for insane things
This removes the delay when switching audio tracks.

Background:

The demuxer packet cache buffers a certain amount of packets. This
includes only packets from selected streams. We discard packets from
other streams for various reasons. This introduces a problem: switching
to a different audio track introduces a delay. The delay is as big as
the demuxer packet cache buffer. This is because the file was read
ahead to fill the packet buffer, which also discarded the packets
from the previously not selected audio stream. Once the remaining
packet buffer has been played, new audio packets are available.

We could probably just not discard packets from unselected streams.
But this would require additional memory and CPU resources, and also
it's hard to tell when packets from unused streams should be discarded
(we don't want to keep them forever; it'd be a memory leak).

[Note: after implementing this code, I haven't thought yet about
whether the discard logic would be easier than this commit.]

So what we do in this commit instead is: we just seek back to the
position where our current packet buffer starts, and start demuxing
from this position again. This way we can get the "past" packets for
the newly selected stream. For streams which were already selected
the packets are simply discarded. We discard up until we reach the
previous position again.

The hard part is not feeding the decoder a packet twice. If we assume
that the demuxer is deterministic (returns exactly the same packets
after a seek to a previous position), then we can try to check whether
it's the same packet as the one at the end of the packet buffer. If
it is, we know that the packet after it is where we left off last time.

Unfortunately, this is not very robust, and maybe it can't be made
robust. Currently we use the demux_packet.pos field as unique packet
ID - which works fine in some naive scenarios, but actually stops
being unique as soon as sub-packets and parsers come into play. These
split multiple smaller packets from a "physical" packet (or even
rewrite these), and will use the same "pos" value as the original
packet. This means the decoder will actually get some packets
multiple times, which may lead to artifacts or decoding errors.

In short, this commit is buggy.