Cocoa: use transparent cursor instead of CGDisplayHideCursor.
On Yosemite (and possibly older versions of OS X), there is a bug in the
cursor state handing where a hidden cursor is not shown when using
nonstandard methods to switch between applications (e.g. a global
hotkey). This bug will cause the cursor to stay hidden until the task
switcher is opened.

This commit sidesteps that issue entirely by simply using a transparent
cursor instead of trying to hide the cursor. This implementation relies
on cursor rects to maintain the cursor state when switching away from
and back to mpv. It also moves the cursor hiding implementation into
MpvEventsView.

Rebased on top of a similar, but slightly inferior patch that got merged
into upstream.